+++ ouroboros-network/sim-tests-lib/Test/Ouroboros/Network/Testnet.hs
@@ -83,7 +83,7 @@ import Ouroboros.Network.PeerSelection.RootPeersDNS.LocalRootPeers
 import Ouroboros.Network.PeerSelection.State.EstablishedPeers qualified as EstablishedPeers
 import Ouroboros.Network.PeerSelection.State.KnownPeers qualified as KnownPeers
 import Ouroboros.Network.PeerSelection.State.LocalRootPeers (HotValency (..),
-           WarmValency (..))
+           LocalRootConfig (..), WarmValency (..))
 import Ouroboros.Network.PeerSelection.State.LocalRootPeers qualified as LocalRootPeers
 import Ouroboros.Network.PeerSelection.Types
 import Ouroboros.Network.PeerSharing (PeerSharingResult (..))
@@ -730,8 +730,8 @@ unit_4177 = prop_inbound_governor_transitions_coverage absNoAttenuation script
               (Script (UseBootstrapPeers [RelayAccessDomain "bootstrap" 00000] :| []))
               (TestAddress (IPAddr (read "0:7:0:7::") 65533))
               PeerSharingDisabled
-              [ (1,1,Map.fromList [(RelayAccessDomain "test2" 65535,(DoNotAdvertisePeer, IsNotTrustable))
-              , (RelayAccessAddress "0:6:0:3:0:6:0:5" 65530,(DoNotAdvertisePeer, IsNotTrustable))])
+              [ (1,1,Map.fromList [(RelayAccessDomain "test2" 65535,LocalRootConfig DoNotAdvertisePeer IsNotTrustable)
+              , (RelayAccessAddress "0:6:0:3:0:6:0:5" 65530,LocalRootConfig DoNotAdvertisePeer IsNotTrustable)])
               ]
               (Script (LedgerPools [] :| []))
               ConsensusModePeerTargets {
@@ -750,11 +750,11 @@ unit_4177 = prop_inbound_governor_transitions_coverage absNoAttenuation script
               False
               (Script (FetchModeDeadline :| []))
           , [JoinNetwork 1.742857142857
-            ,Reconfigure 6.33333333333 [(1,1,Map.fromList [(RelayAccessDomain "test2" 65535,(DoAdvertisePeer, IsNotTrustable))]),
-                                        (1,1,Map.fromList [(RelayAccessAddress "0:6:0:3:0:6:0:5" 65530,(DoAdvertisePeer, IsNotTrustable))
+            ,Reconfigure 6.33333333333 [(1,1,Map.fromList [(RelayAccessDomain "test2" 65535,LocalRootConfig DoAdvertisePeer IsNotTrustable)]),
+                                        (1,1,Map.fromList [(RelayAccessAddress "0:6:0:3:0:6:0:5" 65530,LocalRootConfig DoAdvertisePeer IsNotTrustable)
                                        ])]
-            ,Reconfigure 23.88888888888 [(1,1,Map.empty),(1,1,Map.fromList [(RelayAccessAddress "0:6:0:3:0:6:0:5" 65530,(DoAdvertisePeer, IsNotTrustable))])]
-            ,Reconfigure 4.870967741935 [(1,1,Map.fromList [(RelayAccessDomain "test2" 65535,(DoAdvertisePeer, IsNotTrustable))])]
+            ,Reconfigure 23.88888888888 [(1,1,Map.empty),(1,1,Map.fromList [(RelayAccessAddress "0:6:0:3:0:6:0:5" 65530,LocalRootConfig DoAdvertisePeer IsNotTrustable)])]
+            ,Reconfigure 4.870967741935 [(1,1,Map.fromList [(RelayAccessDomain "test2" 65535,LocalRootConfig DoAdvertisePeer IsNotTrustable)])]
             ]
           )
         , ( NodeArgs 1 InitiatorAndResponderDiffusionMode (Just 135)
@@ -1335,8 +1335,8 @@ unit_4191 = testWithIOSim prop_diffusion_dns_can_recover 125000 absInfo script
             (Script (UseBootstrapPeers [RelayAccessDomain "bootstrap" 00000] :| []))
             (TestAddress (IPAddr (read "0.0.1.236") 65527))
             PeerSharingDisabled
-            [ (2,2,Map.fromList [ (RelayAccessDomain "test2" 15,(DoNotAdvertisePeer, IsNotTrustable))
-                                , (RelayAccessDomain "test3" 4,(DoAdvertisePeer, IsNotTrustable))])
+            [ (2,2,Map.fromList [ (RelayAccessDomain "test2" 15,LocalRootConfig DoNotAdvertisePeer IsNotTrustable)
+                                , (RelayAccessDomain "test3" 4,LocalRootConfig DoAdvertisePeer IsNotTrustable)])
             ]
             (Script (LedgerPools [] :| []))
             ConsensusModePeerTargets {
@@ -1383,10 +1383,10 @@ unit_4191 = testWithIOSim prop_diffusion_dns_can_recover 125000 absInfo script
             , [ JoinNetwork 6.710144927536
               , Kill 7.454545454545
               , JoinNetwork 10.763157894736
-              , Reconfigure 0.415384615384 [(1,1,Map.empty)
+              , Reconfigure 0.415384615384 [(1,1,Map.fromList [])
               , (1,1,Map.empty)]
-              , Reconfigure 15.550561797752 [(1,1,Map.empty)
-              , (1,1,Map.fromList [(RelayAccessDomain "test2" 15,(DoAdvertisePeer, IsNotTrustable))])]
+              , Reconfigure 15.550561797752 [(1,1,Map.fromList [])
+              , (1,1,Map.fromList [(RelayAccessDomain "test2" 15,LocalRootConfig DoAdvertisePeer IsNotTrustable)])]
               , Reconfigure 82.85714285714 []
               ])
         ]
@@ -1470,7 +1470,7 @@ prop_connect_failure (AbsIOError ioerr) =
             naBootstrapPeers = Script (DontUseBootstrapPeers :| []),
             naAddr = TestAddress (IPAddr nodeIP nodePort),
             naPeerSharing = PeerSharingDisabled,
-            naLocalRootPeers = [(1,1,Map.fromList [(RelayAccessAddress relayIP relayPort,(DoNotAdvertisePeer, IsNotTrustable))])],
+            naLocalRootPeers = [(1,1,Map.fromList [(RelayAccessAddress relayIP relayPort,LocalRootConfig DoNotAdvertisePeer IsNotTrustable)])],
             naLedgerPeers = Script (LedgerPools [] :| []),
             naPeerTargets = ConsensusModePeerTargets {
               deadlineTargets = PeerSelectionTargets {
@@ -1602,7 +1602,7 @@ prop_accept_failure (AbsIOError ioerr) =
             naBootstrapPeers = Script (DontUseBootstrapPeers :| []),
             naAddr = TestAddress (IPAddr nodeIP nodePort),
             naPeerSharing = PeerSharingDisabled,
-            naLocalRootPeers = [(1,1,Map.fromList [(RelayAccessAddress relayIP relayPort,(DoNotAdvertisePeer, IsNotTrustable))])],
+            naLocalRootPeers = [(1,1,Map.fromList [(RelayAccessAddress relayIP relayPort,LocalRootConfig DoNotAdvertisePeer IsNotTrustable)])],
             naLedgerPeers = Script (LedgerPools [] :| []),
             naPeerTargets = ConsensusModePeerTargets {
               deadlineTargets = PeerSelectionTargets {
@@ -2549,21 +2549,21 @@ async_demotion_network_script =
           ]
         )
       , ( common { naAddr           = addr2,
-                   naLocalRootPeers = [(1,1, Map.fromList [(ra_addr1, (DoNotAdvertisePeer, IsNotTrustable))])] }
+                   naLocalRootPeers = [(1,1, Map.fromList [(ra_addr1, LocalRootConfig DoNotAdvertisePeer IsNotTrustable)])] }
         , [JoinNetwork 0, Kill 5, JoinNetwork 20]
         )
       , ( common { naAddr           = addr3,
-                   naLocalRootPeers = [(1,1, Map.fromList [(ra_addr1, (DoNotAdvertisePeer, IsNotTrustable))])] }
+                   naLocalRootPeers = [(1,1, Map.fromList [(ra_addr1, LocalRootConfig DoNotAdvertisePeer IsNotTrustable)])] }
         , [JoinNetwork 0]
         )
       ]
   where
     addr1    = TestAddress (IPAddr (read "10.0.0.1") 3000)
     ra_addr1 = RelayAccessAddress (read "10.0.0.1") 3000
-    localRoots1  = [(2,2, Map.fromList [(ra_addr2, (DoNotAdvertisePeer, IsNotTrustable))
-                                       ,(ra_addr3, (DoNotAdvertisePeer, IsNotTrustable))])]
-    localRoots1' = [(2,2, Map.fromList [(ra_addr2, (DoAdvertisePeer, IsNotTrustable))
-                                       ,(ra_addr3, (DoAdvertisePeer, IsNotTrustable))])]
+    localRoots1  = [(2,2, Map.fromList [(ra_addr2, LocalRootConfig DoNotAdvertisePeer IsNotTrustable)
+                                       ,(ra_addr3, LocalRootConfig DoNotAdvertisePeer IsNotTrustable)])]
+    localRoots1' = [(2,2, Map.fromList [(ra_addr2, LocalRootConfig DoAdvertisePeer IsNotTrustable)
+                                       ,(ra_addr3, LocalRootConfig DoAdvertisePeer IsNotTrustable)])]
 
     addr2    = TestAddress (IPAddr (read "10.0.0.2") 3000)
     ra_addr2 = RelayAccessAddress (read "10.0.0.2") 3000
@@ -3129,7 +3129,7 @@ prop_unit_4258 =
              (Script (UseBootstrapPeers [RelayAccessDomain "bootstrap" 00000] :| []))
              (TestAddress (IPAddr (read "0.0.0.4") 9))
              PeerSharingDisabled
-             [(1,1,Map.fromList [(RelayAccessAddress "0.0.0.8" 65531,(DoNotAdvertisePeer, IsNotTrustable))])]
+             [(1,1,Map.fromList [(RelayAccessAddress "0.0.0.8" 65531,LocalRootConfig DoNotAdvertisePeer IsNotTrustable)])]
              (Script (LedgerPools [] :| []))
              ConsensusModePeerTargets {
                deadlineTargets = nullPeerSelectionTargets {
@@ -3164,7 +3164,7 @@ prop_unit_4258 =
              (Script (UseBootstrapPeers [RelayAccessDomain "bootstrap" 00000] :| []))
              (TestAddress (IPAddr (read "0.0.0.8") 65531))
              PeerSharingDisabled
-             [(1,1,Map.fromList [(RelayAccessAddress "0.0.0.4" 9,(DoNotAdvertisePeer, IsNotTrustable))])]
+             [(1,1,Map.fromList [(RelayAccessAddress "0.0.0.4" 9,LocalRootConfig DoNotAdvertisePeer IsNotTrustable)])]
              (Script (LedgerPools [] :| []))
              ConsensusModePeerTargets {
                deadlineTargets = nullPeerSelectionTargets {
@@ -3195,7 +3195,7 @@ prop_unit_4258 =
              False
              (Script (FetchModeDeadline :| []))
          , [ JoinNetwork 3.384615384615,
-             Reconfigure 3.583333333333 [(1,1,Map.fromList [(RelayAccessAddress "0.0.0.4" 9,(DoNotAdvertisePeer, IsNotTrustable))])],
+             Reconfigure 3.583333333333 [(1,1,Map.fromList [(RelayAccessAddress "0.0.0.4" 9,LocalRootConfig DoNotAdvertisePeer IsNotTrustable)])],
              Kill 15.55555555555,
              JoinNetwork 30.53333333333,
              Kill 71.11111111111
@@ -3237,8 +3237,8 @@ prop_unit_reconnect =
               (Script (DontUseBootstrapPeers :| []))
               (TestAddress (IPAddr (read "0.0.0.0") 0))
               PeerSharingDisabled
-              [ (2,2,Map.fromList [ (RelayAccessAddress "0.0.0.1" 0,(DoNotAdvertisePeer, IsNotTrustable))
-                                  , (RelayAccessAddress "0.0.0.2" 0,(DoNotAdvertisePeer, IsNotTrustable))
+              [ (2,2,Map.fromList [ (RelayAccessAddress "0.0.0.1" 0,LocalRootConfig DoNotAdvertisePeer IsNotTrustable)
+                                  , (RelayAccessAddress "0.0.0.2" 0,LocalRootConfig DoNotAdvertisePeer IsNotTrustable)
                                   ])
               ]
               (Script (LedgerPools [] :| []))
@@ -3269,7 +3269,7 @@ prop_unit_reconnect =
                (Script (DontUseBootstrapPeers :| []))
                (TestAddress (IPAddr (read "0.0.0.1") 0))
                PeerSharingDisabled
-               [(1,1,Map.fromList [(RelayAccessAddress "0.0.0.0" 0,(DoNotAdvertisePeer, IsNotTrustable))])]
+               [(1,1,Map.fromList [(RelayAccessAddress "0.0.0.0" 0,LocalRootConfig DoNotAdvertisePeer IsNotTrustable)])]
                (Script (LedgerPools [] :| []))
                ConsensusModePeerTargets {
                  deadlineTargets = PeerSelectionTargets {
@@ -3704,7 +3704,7 @@ unit_peer_sharing =
                (mainnetSimArgs 3)
                (singletonScript (mempty, ShortDelay))
                [ ( (defaultNodeArgs GenesisMode) { naAddr = ip_0,
-                                     naLocalRootPeers = [(1, 1, Map.fromList [(ra_1, (DoNotAdvertisePeer, IsNotTrustable))])],
+                                     naLocalRootPeers = [(1, 1, Map.fromList [(ra_1, LocalRootConfig DoNotAdvertisePeer IsNotTrustable)])],
                                      naPeerTargets = targets 1
                                    }
                  , [JoinNetwork 0]
@@ -3716,7 +3716,7 @@ unit_peer_sharing =
                  , [JoinNetwork 0]
                  )
                , ( (defaultNodeArgs GenesisMode) { naAddr = ip_2,
-                                     naLocalRootPeers = [(1, 1, Map.fromList [(ra_1, (DoNotAdvertisePeer, IsNotTrustable))])],
+                                     naLocalRootPeers = [(1, 1, Map.fromList [(ra_1, LocalRootConfig DoNotAdvertisePeer IsNotTrustable)])],
                                      naPeerTargets = targets 2
                                    }
                  , [JoinNetwork 0]