+++ ouroboros-network/CHANGELOG.md
@@ -8,6 +8,7 @@
 * Removed deprecated `ReconnectDelay` type alias.
 * Addapted to `network-mux` changes in https://github.com/IntersectMBO/ouroboros-network/pull/4999
 * Addapted to `network-mux` changes in https://github.com/IntersectMBO/ouroboros-network/pull/4997
+* Use `LocalRootConfig` instead of a tuple.
 
 ### Non-Breaking changes